Fix k indexing; add test for reverse k order; clarify k.
There seems to have been an inconsistency before. kts:kte is used to declare
array indexing. I believe in that context, Fortran requires kts <= kte. But the
docs said kts was *always* the top index, which conflicts with the first
point. Indeed, a gfortran build, at least, seg faults when kts > kte.

Thus, I added an explicit kdir argument, and doc'ed it.

With both directions now working, I added a test so that now the same ICs are run twice, the second time with indexing the opposite of the first.