Rm -fsingle-precision-constant flag.
Turns out this flag causes compile erros in the STL, apparent once I included a
nontrivial header like Kokkos_Core.hpp. Example:

/projects/sems/install/rhel7-x86_64/sems/compiler/gcc/5.3.0/base/include/c++/5.3.0/bits/random.tcc:1408:55: error: no matching function for call to ‘max(float, const double&)’
    _M_d = std::round(std::max(6.0, std::min(__m, __dx)));

Here the issue is 6.0 is assumed to be double in this particular STL impl. Thus,
we need to gain BFB by another means. A follow-up commit will focus on that.