Skeleton of a nano-app to study vectorization.
This nano-app looks at a stripped-down sedimentation loop with explicit-Euler
first-order upwind.

unittest() checks that all impls are right.

measure_perf() looks at various impls.

namespace ko has a true Kokkos impl that runs on GPU. All other namespaces are
meant to study trade offs relative to this true Kokkos impl.

namespace vec2 has the start of explicit use of intrinsics. So far, on SKX, it
appears that auto-vectorization is nearly perfect, so use of intrinsics has yet
to yield speedup. In addition, it appears that 256-bit-wide vectors can be
better than 512-bit-wide on SKX. Have not checked performance on KNL yet.

Next step is to run with multiple columns.