Add kopack, a combination of the true Kokkos impl and Pack.
kopack demonstrates the pattern we expect, based on previous experience, to use
in P3. It uses the team policy, best for GPU, and packs, best for AVX-type
architectures. It does not use AVX intrinsics. It gets almost as good of
performance as packsimd on KNL and SKX, and almost as good of performance as ko
on GPU.