Config and regression: If DEBUG, do things to make KNL testable.
We want clean ctests on all platforms. On some, with optimized
builds, that's impossible. If building DEBUG, disable
vectorization, and enable FPE.

Fix an issue in the regression tester. We require BFB in double
precision debug builds. The tol specification was erroneously
altered to allow non-BFB (with, in fact, a very loose tolerance)
in this special build. Insert a sanity check to protect this
requirement in the future. Remove the convenient TOL b/c we want
it to be inconvenient to request other than BFB.