Compiler workaround.
The perf drivers weren't running correctly in a build with Intel
17 on KNL, opt flags. Turns out it's b/c the ICs weren't filling
correctly b/c of an error in the dz arrays. This error was a
result of a compiler mistranslation. I used the usual technique
of letting the compiler know it overlooked a loop dependency:
adding volatile to the variable that causes the dependency.