Work on Kokkos inter-column ||ism.
1. Fuse the 1/dzq computation so there is only one top-level || region.

2. Fix concurrency bugs of two types: missing barriers, and a write race
condition on accumulation of a quantity.

3. In debug builds, use 7 threads, if available, per team. On non-GPU platforms
with OpenMP enabled and ctest run as, say,
    OMP_NUM_THREADS=17 ctest,
we'll get testing of these kinds of bugs. Often OpenMP is far more effective at
fishing these out than the GPU, as the GPU has h/w-related syncs that can hide
bugs. On GPU, the error due to the missing barriers is hidden by the fact that
AUTO is making a 1-warp team (bad: I always recommend against using AUTO). I'm
not yet sure why we didn't see the accumulation bug on GPU, but with OpenMP and
inter-column ||ism, it's immediately obvious.

Number 3 is a general tactic for testing fine ||ism: even it's meant to be used
in practice only on GPU (which has ~10^2 more ||ism than anything else), we can
test correctness anywhere using OpenMP. We'll build this tactic into all SCREAM
development work.