Comment out some unused calcs.
I found that the optimizer was usually, but not always, removing these calcs. It
was messing up a bit of analysis. Comment these out for now. If we want them in
later as a model of intensive flops, we can force them to be computed by any of
several means.