Handle parallel_reduce corner case.
If qr is everywhere too small, exit the dt_left loop. This check is not
necessary if the Co_max_local calc is done regardless of qr_gt_small.any(). But
that's a waste of work. Because Kokkos::parallel_reduce exits with Co_max set to
Max::init() in this case, we have to explicitly handle it, since Co_max < 0
breaks the code in the remainder of the loop.