Packize find_lookupTable_indices_3_kokkos, but don't accept it.
This is a case where we might be better with the pack -> scalar wrapper. KNL
performance is worse with the pack impl than the scalar (but it's better with
the top-level pack -> scalar impl than a full-scalar impl).

For now, ifdef it out. The impl is instructional b/c it handles more complicated
masking and also dealing with FPEs when using packs.

We'll come back to this later to see if the pack impl can be improved.