Pack tweaks.
Remove an unused function. Masked loops should be handled through a macro, not a
loop over a functor, for at least two reasons:
  1. Easy to trace vectorization messages to actual source, rather than just to
     the same 'loop' function.
  2. Vectorization is not assured to occur over functor calls.

Switch | to || and & to &&. Originally I was capturing the idea of a bit
mask. But after thining about it, I decided it's more consistent with the rest
of the ops that a scalar op should simply apply to a pack.