Tweaks to Workspace(Manager).
1. Clean up some arithmetic. Some of this is according to taste, and some is
objectively a little cleaner. The main thing is there is now only integer
arithmetic and no casts to real type or calls to ceil. A second thing is I stuck
in m_total and make m_size the user's size. This simplifies (IMO) some of the
arithmetic.

2. Use a wrapper to release_impl to help the compiler deduce a template type
that is in principle deducible but is not necessarily deducible in the C++
standard. Use FORCEINLINE to make this one-line wrapper cost nothing.

3. Move singles into take and release_impl according to our discussion. The idea
here is that workspace metadata is manipulated by a single thread, but all
threads must create their own View (metadata) wrapping the global bulk
data. From a user's perspective, a side effect is that take and release are bare
function calls w/o need of being in a single region. (There is still a single
region in the unit test, but that's for correctness checking, not for
take/release calls.)

4. Rework the team_barriers a bit to support 3.

5. In the unit test, remove the global set of views outside of the
kernel. Replace with a static array of Views decl'ed in the kernel, mirroring
intended convenient Workspace usage.