Fix a bug in WorkspaceManager::init().
init() was calling get_workspace in a kernel that uses a
different policy than the WM client's policy. That is invalid, as
the map of threads to teams will in general be different.

Separately, choose just one of concurrency() and
omp_get_max_threads() as the means to get the total thread
count. Previously, we were using one or the other
inconsistently. In principle they return the same value, but a
comment in the Kokkos code suggests that omp_get_max_threads()
can have side effects. Since Kokkos will end up calling
concurrency() internally, let's use that one.

Also move some function defs to util.cpp, just as a cleanup thing
I noticed.