Fix a bug in upwind test. Add more upwind documentation.
The unit test can't use a restricted-domain call for r[0] in the first time
step, but we want to test the restricted-domain call. Thus, in the first time
step, make two separate calls for r[0] (full domain) and r[1] (restricted
domain). This fixes the bug that the top cell's rho value was ignored when
advecting rho. (The extrema test didn't fail b/c the neglect of this value has
little impact. In contrast, the extrema test *does* fail if tracer consistency
is not enforced.)

Add a single-mixing ratio upwind wrapper and test it.

Add more documentation, including on the issue of const and in nested data
structures.