Document a possible bug in find_lookupTable_indices_3-related interpolation.
The factor inv_dum3 in the interpolation formula leads to a discontinuous
interpolant. Removal of it gives a smooth interpolant, and the resulting formula
is also analytically a bilinear interpolant. This commit starts a unit test for
the ...indices_3-related code and in particular writes data to draw a slice in
the (1 + mu_r)/lamr direction for fixed mu_r. Subsequent commits will replace
this code with tests of continuity and safety bounds.