test-all-scream: Generate baselines from master
test-all-scream now accepts the --baseline argument which allows
the user to specific which branch/sha should be used to generate
baselines; this value defaults to origin/master.

test-all-scream now has considerable extra code that deals with
switch the users' repo over to the baseline sha, generating the baselines,
and swithing it back.