Tridiag nano-app updates.
Probably the final version of the monolithic Makefile-based nano-app.

I'm thinking I might keep this as is in the scream-docs repo, b/c this style of
code and build system makes it easy to be sure everything is working as expected
on a machine. Then I might do the rest of the development/docs in the scream
repo directly, with the unit test exe also serving as a performance driver. This
approach will let me verify I'm getting the expected performance when integrated
into the scream repo with the cmake build system.